gcc 6-size.c -m32 -o size32 2> /tmp/32

+++ 0x00-hello_world/6-size.c
@@ -7,11 +7,11 @@
 
 int main(void)
 {
-	printf("Size of a char: %lu\n", sizeof(char));
-	printf("Size of an int: %lu\n", sizeof(int));
-	printf("Size of a long int: %lu\n", sizeof(long int));
-	printf("Size of a long long int: %lu\n", sizeof(long long int);
-	printf("Size of a float: %lu\n", sizeof(float));
+	printf("Size of a char: %lu\n", (unsigned long)sizeof(char));
+	printf("Size of an int: %lu\n", (unsigned long)sizeof(int));
+	printf("Size of a long int: %lu\n", (unsigned long)sizeof(long int));
+	printf("Size of a long long int: %lu\n", (unsigned long)sizeof(long long int);
+	printf("Size of a float: %lu\n", (unsigned long)sizeof(float));
 
 	return (0);
 }
